+++ js/check-line-length.js
@@ -1,0 +1,3 @@
+const checkMaxLength = (str, max = 100) => str.length <= max;
+
+export{checkMaxLength};
