+++ js/get-random-number.js
@@ -1,0 +1,12 @@
+const getRandomInt = (min, max) => {
+    if(min < 0 || max < 0) {
+      return -1;
+    }
+    if(min > max) {
+      [min, max] = [max, min];
+    }
+    return Math.floor(Math.random() * (max - min + 1)) + min;
+  };
+
+  export{getRandomInt};
+
